provider: send an explicitly configured temperature of 0 instead of dropping it
CompletionRequest.Temperature was a float64, so its zero value (unset) was
indistinguishable from an explicit 0 and every provider's Complete guarded on
`req.Temperature != 0`. Three paths configure temperature 0 and document it as
deterministic (the enrichment ingestion decision, the rerank judge's
per-candidate scoring, and ask's sub-query decomposition), so all three were
silently sampled at the model's server default rather than decoding greedily.

Make Temperature a *float64: nil omits the field (server default, unchanged for
callers that never set it) and a non-nil value is sent verbatim, including 0.
Add a Float64 helper, wrap the request-construction sites, and update the three
provider guards to a nil check. The ingestion retry-skip now compares the
resolved temperature it owns rather than reading it back out of the request.

Bump to 0.18.45.

+++ internal/provider/reranker.go
@@ -233,7 +233,7 @@ func (j *judgeReranker) Rerank(ctx context.Context, query string, docs []string)
 		user := Fence("query", query) + "\n\n" + Fence("document", doc)
 		resp, err := j.llm.Complete(ctx, &CompletionRequest{
 			Messages:    BuildGuardedMessages(jc.SystemPrompt, user),
-			Temperature: jc.Temperature,
+			Temperature: Float64(jc.Temperature),
 			MaxTokens:   jc.MaxTokens,
 		})
 		if err != nil {
